Reorder directories
