+++ examples/partitioned-heat-conduction-IGA-Dirichlet.cpp
