+++ examples/partitioned-heat-conduction-IGA-Neumann.cpp
