--- examples/partitioned-heat-conduction-direct/gismo-left/run.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -e -u
-
-../../../../../build/bin/partitioned-heat-conduction -c ../precice-config.xml --plot -s 0
