--- examples/partitioned-heat-conduction-vertex-vertex/gismo-left/clean.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-set -e -u
-
-. ../../tools/cleaning-tools.sh
-
-clean_gismo .
