+++ examples/partitioned-heat-conduction/partitioned-heat-conduction-IGA-IGA/dirichlet-gismo/run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e -u
+
+../gismo-executable/gismo-executable-dirichlet -c ../precice-config.xml
