+++ examples/partitioned-heat-conduction/partitioned-heat-conduction-IGA-IGA/precice-config.xml
