+++ examples/partitioned-heat-conduction/partitioned-heat-conduction-vertex-vertex/dirichlet-gismo/run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e -u
+
+../gismo-executable/gismo-executable -c ../precice-config.xml --plot -s 0
