--- examples/perpendicular-flap-IGA-IGA/precice-config.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<precice-configuration>
-  <log>
-    <sink
-      filter="%Severity% > info and %Rank% = 0"
-      format="---[precice] %ColorizedSeverity% %Message%"
-      enabled="false" />
-
-    <sink
-      filter="%Severity% > debug and %Rank% = 0"
-      format="---[precice] %ColorizedSeverity% %Message%"
-      enabled="true" />
-  </log>
-
-  <data:vector name="GeometryControlPointData" />
-  <data:vector name="ForceControlPointData" />
-  <data:vector name="GeometryKnotData" />
-  <data:vector name="ForceKnotMeshData" />
-
-
-  <mesh name="GeometryKnotMesh" dimensions="2">
-    <use-data name="GeometryKnotData" />
-  </mesh>
-  
-  <mesh name="GeometryControlPointMesh" dimensions="3">
-      <use-data name="GeometryControlPointData" />
-  </mesh>
-  
-  <mesh name="ForceKnotMesh" dimensions="2">
-      <use-data name="ForceKnotMeshData" />
-  </mesh>
-  
-  <mesh name="ForceControlPointMesh" dimensions="3">
-      <use-data name="ForceControlPointData" />
-  </mesh>
-  
-  <participant name="Fluid">
-      <provide-mesh name="ForceControlPointMesh" />
-      <provide-mesh name="ForceKnotMesh" />
-      <receive-mesh name="GeometryControlPointMesh" from="Solid" direct-access="true"/>
-      <receive-mesh name="GeometryKnotMesh" from="Solid" />
-      <read-data    name="GeometryControlPointData" mesh="GeometryControlPointMesh" />
-<!--      <read-data    name="GeometryKnotData" mesh="GeometryKnotMesh" />-->
-      <write-data   name="ForceControlPointData" mesh="ForceControlPointMesh" />
-<!--      <write-data   name="ForceKnotMeshData" mesh="ForceKnotMesh" />-->
-  </participant>
-  
-  <participant name="Solid">
-      <provide-mesh name="GeometryControlPointMesh" />
-      <provide-mesh name="GeometryKnotMesh" />
-      <receive-mesh name="ForceControlPointMesh" from="Fluid" direct-access="true"/>
-      <receive-mesh name="ForceKnotMesh" from="Fluid" />
-<!--      <read-data    name="ForceKnotMeshData" mesh="ForceKnotMesh" />-->
-      <read-data    name="ForceControlPointData"            mesh="ForceControlPointMesh" />
-      <write-data   name="GeometryControlPointData" mesh="GeometryControlPointMesh" />
-<!--      <write-data   name="GeometryKnotData" mesh="GeometryKnotMesh" />-->
-  </participant>
-  
-  <m2n:sockets acceptor="Fluid" connector="Solid" exchange-directory="../"/>
-  <coupling-scheme:serial-implicit>
-      <participants first="Solid" second="Fluid" />
-      <max-time value="1" />
-      <time-window-size value="0.01" />
-      <exchange data="GeometryControlPointData" mesh="GeometryControlPointMesh" from="Solid" to="Fluid" />
-<!--      <exchange data="GeometryKnotData" mesh="GeometryKnotMesh" from="Solid" to="Fluid" />-->
-      <exchange data="ForceControlPointData" mesh="ForceControlPointMesh" from="Fluid" to="Solid" />
-<!--      <exchange data="ForceKnotMeshData" mesh="ForceKnotMesh" from="Fluid" to="Solid" />-->
-      <max-iterations value="50" />
-      <relative-convergence-measure limit="1e-4" data="GeometryControlPointData"  mesh="GeometryControlPointMesh"/>
-      <relative-convergence-measure limit="1e-4" data="ForceControlPointData"         mesh="ForceControlPointMesh"/>
-      <acceleration:IQN-ILS>
-<!--        <data name="GeometryControlPointData" mesh="GeometryControlPointMesh" />-->
-        <data name="ForceControlPointData"        mesh="ForceControlPointMesh" />
-        <preconditioner type="residual-sum" />
-        <filter type="QR2" limit="1e-4" />
-        <initial-relaxation value="1.0" />
-        <max-used-iterations value="100" />
-        <time-windows-reused value="15" />
-      </acceleration:IQN-ILS>
-      <acceleration:constant>
-        <relaxation value="1.0" />
-      </acceleration:constant>
-    </coupling-scheme:serial-implicit>
-</precice-configuration>
